+++ game-mewgenics/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - None Planned
 
+## [0.2.0] - 2026-04-22
+
+- Improved: Load Order rendering using React - show mod image thumbnails
+- Fixed: Properly assign mod attribute at install for Load Order
+
+## [0.1.3] - 2026-03-25
+
+- Added: Mewjector support
+
 ## [0.1.2] - 2026-03-25
 
 - Added: Mewjector support